requirement.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
idna==3.3
joblib==1.1.0
nltk==3.7
PySastrawi==1.2.0
PySide6-Essentials==6.3.1
requests==2.28.1
shiboken6==6.3.1
tomli==2.0.1
tqdm==4.64.0
urllib3==1.26.9
wn==0.9.1
